Lint scss/js
